Time: 30 ms (6.58%), Space: 8.6 MB (50.23%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -9,7 +9,7 @@ public:
         {
             return 2;
         }
-        for(double i=0;i<x/2;i++)
+        for(long i=0;i<x/2;i++)
         {
             if(x==i*i)
             {
